refactor(database): move interface into repository
